# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Always_date_api documentation",
        "description": "Always date documentation for back-end",
        "contact": {
            "name": "FlokY#0101"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Lietotaja izrakstīšanās",
                "description": "Lietotaja izrakstīšanās no konta",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Veiksmīga izrakstīšanās.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/life_school_comment": {
            "get": {
                "tags": [
                    "Life school comment"
                ],
                "summary": "Iegūst visus dzīves skolas komentārus",
                "description": "Iegūst visus dzīves skolas komentārus",
                "operationId": "getLifeSchoolComment",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolCommentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Life school comment"
                ],
                "summary": "Tiek uzrakstīts un saglabāts komentārs dzīves skolas rakstam",
                "description": "Ievada komentāru zem kādas dzīves skolas raksta, komentārs glabājas un, glabā visus nepiciešamos datus, lai būtu saistīts ar konkrētu komentāru.",
                "operationId": "postLifeSchoolComment",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#components/schemas/LifeSchoolCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolCommentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/life_school_comment/{id}": {
            "get": {
                "tags": [
                    "Life school comment"
                ],
                "summary": "Iegūst konkrētu komentāru",
                "description": "Iegūst konkrētu komentāru, ar tā komentāra palīdzību",
                "operationId": "getLifeSchoolCommentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "LifeSchoolComment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolCommentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Life school comment"
                ],
                "summary": "Atjauno vai reģidē komentāru",
                "description": "Atjauno vai reģidē komentāru pēc id",
                "operationId": "updateLifeSchoolComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "LifeSchoolComment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#components/schemas/LifeSchoolCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolCommentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Life school comment"
                ],
                "summary": "Izdzēš komentāru",
                "description": "Dzēš komentāru pēc id",
                "operationId": "deleteLifeSchoolComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "LifeSchoolComment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolCommentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/life_school_comment/rate": {
            "post": {
                "tags": [
                    "Life school comment"
                ],
                "summary": "Komentāru vērtēšana",
                "description": "Lietotās spēj novērtēt komentāru ar like vai dislike. Viss tiek validēts tā, lai lietotājs nevar novērtēt komentru ar nezināmu vērtējumu vai ar 2 vērtējumiem vienlaicīgi",
                "operationId": "rateLifeSchoolComment",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#components/schemas/CommentRatingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolCommentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nevar novērtēt komentāru ar nezinānu vērtējumu."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LifeSchoolCommentResource": {
                "title": "Life school comment resource",
                "description": "Life School comment resource",
                "xml": {
                    "name": "LifeSchoolCommentResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}